drf_api/requirements.txt
rapidfuzz==2.11.1
amqp
appdirs==1.4.4
asgiref==3.7.2
attrs==22.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach
boto3==1.26.2
botocore==1.29.2
cached-property==1.5.2
cachetools==5.2.0
celery==5.2.7
certifi
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
conf==0.4.1
cryptography==38.0.4
dataclasses
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
Django==3.2.15
django-auditlog==1.0.0
django-colorfield==0.7.2
django-jsonfield-backport==1.0.5
django-pglocks==1.0.4
djangorestframework==3.14.0
drf-api-logger==1.0.6
elasticsearch==8.4.0
et-xmlfile==1.1.0
filelock==3.9.0
flower==1.0.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.60.0
google-auth==2.17.3
google-auth-httplib2==0.1.1
google-auth-oauthlib
googleapis-common-protos==1.56.3
gspread==5.10.0
httplib2==0.22.0
humanfriendly==10.0
humanize==4.2.0
idna==3.4
imbox==0.9.8
importlib-metadata==4.11.4
ipython==7.34.0
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.18.2
jmespath==1.0.0
kombu
lxml==4.9.2
Naked==0.1.31
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.1
pandas
paramiko==2.9.2
parso==0.8.3
pem==21.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.22.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
Pygments==2.13.0
pymongo==4.3.3
PyNaCl==1.5.0
pyparsing==3.0.9
PyMuPDF==1.18.19
pypdf==3.3.0
pysftp==0.2.9
python-dateutil==2.8.2
python-logstash==0.4.6
python-redis-lock==3.7.0
pytz==2023.3
PyYAML==6.0
redis==4.5.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
s3transfer==0.6.0
selenium==4.10.0
sentry-sdk==2.14.0
shellescape==3.8.1
six==1.16.0
slack-sdk==3.21.1
soupsieve==2.4.1
sqlparse==0.4.3
tabula==1.0.5
tabula-py==2.3.0
timeago==1.0.15
tornado==6.2
traitlets==5.7.1
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.13
utils==1.0.1
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.0.3
xmltodict==0.12.0
xmltojson==0.2.0
zeep==4.0.0
zipp==3.12.0
